% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_avg.R
\name{compute_avg}
\alias{compute_avg}
\title{Compute average characteristics for identified fruit}
\usage{
compute_avg(mini_dat)
}
\arguments{
\item{mini_dat}{a data frame}
}
\value{
a named numeric vector
}
\description{
Compute average characteristics for identified fruit
}
\examples{
compute_avg(fruit)
}
