% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit_avg.R
\name{find_cols}
\alias{find_cols}
\title{Find columns matching a pattern}
\usage{
find_cols(dat, pattern)
}
\arguments{
\item{dat}{a data frame}

\item{pattern}{character string of pattern to identify}
}
\value{
a numeric vector of column positions
}
\description{
Find columns matching a pattern
}
\examples{
find_cols(fruit, "berry")
}
