% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit.R
\docType{data}
\name{fruit}
\alias{fruit}
\title{Fruit characteristics}
\format{
A data frame with 3 rows and 4 variables:
\describe{
\item{blackberry}{blackberry calories, weight, and yumminess}
\item{blueberry}{blueberry calories, weight, and yumminess}
\item{peach}{peach calories, weight, and yumminess}
\item{plum}{plum calories, weight, and yumminess}
}
}
\source{
\url{https://github.com/jennybc/debugging}
}
\usage{
fruit
}
\description{
A dataset containing calories, weight, and yumminess (rows) of different
kids of fruit (columns)
}
\keyword{datasets}
