% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit_avg.R
\name{fruit_avg}
\alias{fruit_avg}
\title{Compute average characteristics for identified fruit}
\usage{
fruit_avg(dat, pattern)
}
\arguments{
\item{dat}{a data frame}

\item{pattern}{character string of pattern to identify fruits of interest}
}
\value{
a named numeric vector with average characteristics for identified fruit
}
\description{
Compute average characteristics for identified fruit
}
\examples{
fruit_avg(fruit, "berry")
}
