% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_fruit.R
\name{message_fruit}
\alias{message_fruit}
\title{Message user about how many fruits are found (i.e, how many
columns are in the data).}
\usage{
message_fruit(mini_dat)
}
\arguments{
\item{mini_dat}{a data frame}
}
\value{
a message in console
}
\description{
Message user about how many fruits are found (i.e, how many
columns are in the data).
}
\examples{
message_fruit(fruit)
}
