% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit_avg.R
\name{reduce_dat}
\alias{reduce_dat}
\title{Reduce data to only columns of interest}
\usage{
reduce_dat(dat, cols)
}
\arguments{
\item{dat}{a data frame}

\item{cols}{a numeric vector of column positions}
}
\value{
a data frame
}
\description{
Reduce data to only columns of interest
}
\examples{
reduce_dat(fruit, 1:2)
}
